requirements.txt
tensorflow
keras
pandas
hyperopt
optunity
imblearn
ConfigSpace
hpbandster
tqdm
hpsklearn
scikit-learn
numpy
matplotlib
seaborn
mlflow
dvc
dotenv